split wave legacy parsing, add debug texture map
also fixed a very old bug in drs parsing that i didn't see cuz i'm
testing my code on little endian machines only...
