+++ game/src/engine/legacy.cpp
@@ -48,8 +48,8 @@ DRS::DRS(const std::string &path) : in(path, std::ios_base::binary), items() {
 	DrsHdr hdr{ 0 };
 	in.read((char*)&hdr, sizeof(hdr));
 
-	hdr.nlist = le16toh(hdr.nlist);
-	hdr.listend = le16toh(hdr.listend);
+	hdr.nlist = le32toh(hdr.nlist);
+	hdr.listend = le32toh(hdr.listend);
 
 	//printf("%40s %16s\nnlist=%u, listend=%u\n", hdr.copyright, hdr.version, hdr.nlist, hdr.listend);
 
@@ -84,24 +84,6 @@ DRS::DRS(const std::string &path) : in(path, std::ios_base::binary), items() {
 	}
 
 	printf("items read: %llu\n", (unsigned long long)items.size());
-}
-
-std::vector<uint8_t> DRS::open_wav(DrsId k) {
-	ZoneScoped;
-	uint32_t id = (uint32_t)k;
-	DrsItem key{ id, 0, 0 };
-	auto it = items.find(key);
-	if (it == items.end())
-		throw std::runtime_error(std::string("invalid audio ID: ") + std::to_string(id));
-
-	const DrsItem &item = *it;
-
-	// fetch data
-	in.seekg(item.offset);
-	std::vector<uint8_t> data(item.size, 0);
-	in.read((char*)data.data(), item.size);
-
-	return data;
 }
 
 Slp DRS::open_slp(DrsId k) {
